fix: 깨진 테스트 수정
