+++ build.gradle
@@ -6,7 +6,7 @@ plugins {
 
 sourceCompatibility = '11'
 
-allprojects {
+subprojects {
     apply plugin: 'java'
     apply plugin: 'org.springframework.boot'
     apply plugin: 'io.spring.dependency-management'
