--- console/src/test/java/racingcar/RacingCarApplicationTests.java
@@ -1,12 +1,0 @@
-package racingcar;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class RacingCarApplicationTests {
-
-    @Test
-    void contextLoads() {
-    }
-}
