+++ domain-h2/src/test/java/racingcar/TestApplication.java
@@ -1,0 +1,8 @@
+package racingcar;
+
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class TestApplication {
+
+}
